server/requirements.txt
Django == 3.0.4
djangorestframework == 3.11.0
djangorestframework-jwt == 1.11.0
python-jose == 3.1.0
cryptography == 2.8
requests~=2.22.0
pyjwt~=1.7.1
django-cors-headers~=3.1.1
djangorestframework-simplejwt==4.4.0